Renaming of manager modules and first use of module "Event_manager"
