+++ helpers/eve_man.py
@@ -13,16 +13,25 @@ class Event_man():
     def add_timed_event(self, event_function, duration):
         """ Add a time based event provided as "event_function" that will be triggered after "duration" is expired """
         self._timed_events.append({"event": event_function, "trigger_time": self.get_trigger_time(duration)})
-        logging.debug(f"New timed event added: {self._timed_events[-1]}")
+        logging.debug(f"New timed event with duration of {duration}s added: {self._timed_events[-1]['event'].__class__}.{self._timed_events[-1]['event'].__name__}")
 
     def get_trigger_time(self, duration):
         """ Calculate the target trigger time based on the given duration """
-        trigger_time = time.time() + duration
+        trigger_time = time.time() + float(duration)
         return trigger_time
 
     def check_events(self):
         """ Execute stored event function if corresponting "trigger_time" is reached / exceeded """
         current_time = time.time()
+        executed_events = []
+
         for event in self._timed_events:
-            if event["trigger_time"] >= current_time:
-                eval(event["event"])
+            # Execute all events and register them as executed
+            if event["trigger_time"] <= current_time:
+                event["event"]()
+                executed_events.append(event)
+                logging.warning(f"Timed event triggered, trigger time was{time.ctime(event['trigger_time'])}")
+
+        for executed_event in executed_events:
+            # Remove all executed events from the event register
+            self._timed_events.remove(executed_event)
